refactor: 인터페이스 제거 및 클래스명 변경, 패키지 분리
