+++ backend/src/main/java/com/woowacourse/momo/group/service/GroupSearchService.java
@@ -11,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 
 import com.woowacourse.momo.group.domain.Group;
-import com.woowacourse.momo.group.domain.GroupRepository;
+import com.woowacourse.momo.group.infrastructure.querydsl.GroupFindRepository;
 import com.woowacourse.momo.group.service.dto.request.GroupFindRequest;
 import com.woowacourse.momo.group.service.dto.response.GroupPageResponse;
 import com.woowacourse.momo.group.service.dto.response.GroupResponse;
@@ -29,7 +29,7 @@ public class GroupSearchService {
 
     private final MemberFindService memberFindService;
     private final GroupFindService groupFindService;
-    private final GroupRepository groupRepository;
+    private final GroupFindRepository groupFindRepository;
 
     public GroupResponse findGroup(Long id) {
         Group group = groupFindService.findGroup(id);
@@ -38,7 +38,7 @@ public class GroupSearchService {
 
     public GroupPageResponse findGroups(GroupFindRequest request) {
         Pageable pageable = PageRequest.of(request.getPage(), DEFAULT_PAGE_SIZE);
-        Page<Group> groups = groupRepository.findGroups(request.toFindCondition(), pageable);
+        Page<Group> groups = groupFindRepository.findGroups(request.toFindCondition(), pageable);
         List<Group> groupsOfPage = groups.getContent();
         List<GroupSummaryResponse> summaries = GroupResponseAssembler.groupSummaryResponses(groupsOfPage);
 
@@ -48,7 +48,7 @@ public class GroupSearchService {
     public GroupPageResponse findParticipatedGroups(GroupFindRequest request, Long memberId) {
         Pageable pageable = PageRequest.of(request.getPage(), DEFAULT_PAGE_SIZE);
         Member member = memberFindService.findMember(memberId);
-        Page<Group> groups = groupRepository.findParticipatedGroups(request.toFindCondition(), member, pageable);
+        Page<Group> groups = groupFindRepository.findParticipatedGroups(request.toFindCondition(), member, pageable);
         List<Group> groupsOfPage = groups.getContent();
         List<GroupSummaryResponse> summaries = GroupResponseAssembler.groupSummaryResponses(groupsOfPage);
 
@@ -58,7 +58,7 @@ public class GroupSearchService {
     public GroupPageResponse findHostedGroups(GroupFindRequest request, Long memberId) {
         Pageable pageable = PageRequest.of(request.getPage(), DEFAULT_PAGE_SIZE);
         Member member = memberFindService.findMember(memberId);
-        Page<Group> groups = groupRepository.findHostedGroups(request.toFindCondition(), member, pageable);
+        Page<Group> groups = groupFindRepository.findHostedGroups(request.toFindCondition(), member, pageable);
         List<Group> groupsOfPage = groups.getContent();
         List<GroupSummaryResponse> summaries = GroupResponseAssembler.groupSummaryResponses(groupsOfPage);
